feat: implement Node struct and ChannelMessage enum
`Node` is a p2p (network) node (contains server and single client), it
contact with the UI through channels.

Messages sent through these channels are dictated by `ChannelMessage`
enum.

`Node` has a new function, which recievers a reciever end of the
channel from the UI part and returns `Node` and a reciever for the UI part.